v0.2.7 - Resource and CP rework + several menu and game ui improvements

+++ EQFleet/core/src/main/java/de/instinct/eqfleet/game/frontend/projectile/ProjectileInstance.java
@@ -12,7 +12,7 @@ public class ProjectileInstance {
 	
 	private int projectileId;
 	private Vector2 lastPosition;
-	private float aoeRadius;
+	private double aoeRadius;
 	private ModelInstance model;
 	private String particlesTag;
 	private boolean active;
